gdtoa: Fix gdtoa's internal allocator on CHERI
The allocator incorrectly assumed that doubles have equal or stronger
alignment constraints than pointers.

Double the size of the static buffer to account for larger pointers.
This is probably too much, but no mechanism is provided to compute the
optimal value and the comments are quite equivocal.

Effort:		CHERI upstreaming
Sponsored by:	DARPA, AFRL, Innovate UK