tsearch: use size_t rather than uintptr_t to store bits
This fixes the computation of STEPS_BIT (which conflates sizeof()
with the number of bits an integer type can store in a way that
doesn't hold for uintptr_t on CHERI).  It also saves a very small
amount of space on the stack of tsearch/tdelete in the CHERI case.

Effort:		CHERI upstreaming
Sponsored by:	DARPA, AFRL, Innovate UK