CHERI: opportunistically set bounds on pointers returned by fgetwln()
fgetwln() returns a pointer into the buffer of the I/O stream which
is valid until the next I/O operation on the stream.

Because len may be arbitrary, these bounds may not be precise.
Further, the underlying allocation'slifetime is more stack-like than
heap-like in that reuse should be common and rapid.  This means that
CHERI protections are advsisory at best and fgetwln should be avoided
in secure software.

Effort:		CHERI upstreaming
Sponsored by:	DARPA, AFRL, Innovate UK