sys/assym: use 8, not 16, bits when smuggling values
This is a particularly... unusual dance and it breaks on CHERI.

`sys/assym.h` is used to create object files whose symbols have
structured names and associated values which encode constants available
to the C compiler, for example, the evaluation results of offsetof()
expressions.  Symbols and values can be extracted using `nm`.
`genoffset.sh` contains a pile of awk which inverts the encoding and
generates a C file which can be used to verify that the export was
successful.  `genassym.sh` contains a very similar pile of awk to
generate a #define-ful header file for use with C-preprocessing
assemblers.

The encoding that used to be used was to export 16 bits of the magnitude
of the value at a time, creating four (...{`w0`, `w1`, `w2`, `w3`})
symbols per export, as well as a symbol encoding the sign bit.  The
value of each symbol is, in particular, expressed in C as the length of
a global array.  In order to avoid zero-length arrays, an `ASSYM_BIAS`
value is used; to facilitate later (textual!) processing, this was
chosen to be `0x10000`: masking by `0xFFFF` or, equivalently, slicing
off the last four hex characters of the value suffices to recover the
component of the encoding.

Unfortunately, because the symbols' values are not their length, but
rather an *address*, compiler changes which vary global layout will
break this dance.  CHERI, for example, will pad globals so that no other
globals will occur within the most narrow representable bounds for the
object.  This means that attempting to encode 0x158, for example, which
results in an attempt to assign the value 0x10158, instead results in
0x10180.

To work around the issue, switch to encoding byte-by-byte: construct 8
symbols, one for each byte, using an `ASSYM_BIAS` of 0x100.
Correspondingly, adjust the awk script to use two, not four, characters.
This will break if ever CHERI cannot exactly represent objects whose
sizes are in [256, 512).

Effort:		CHERI upstreaming
Sponsored by:	DARPA, AFRL