kevent: avoid smuggling kernel pointers via kevent.data
The AIO subsystem uses the kevent.data field to pass the aio and lio
job pointers to the event attach filter.  This breaks on CHERI because
kevent.data and knote.kn_sdata are not meant to hold pointers.  We add a
separate argument to kqfd_register and kqueue_register to explicitly set
the knote.kn_ptr value upon registration.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK