efirt: Shim AArch64 interface for pure-capability Morello kernels
EFI remains totally oblivious of CHERI so we only get integer-based
tables and functions. We therefore need to mess with the members of the
structs to match the legacy AArch64 layout, derive capabilities from the
integers ourselves and shim calls to follow the legacy AArch64 calling
convention, including with an almighty DDC and making sure LR has the
LSB clear so the firmware's RET to X30 doesn't fault due to the LSB
unaligning the address.