rtld: Support multiple PLTs
Add a new Plt_Entry structure that describes the data associated with
a PLT including the ELF relocation table and PLT GOT.  Count the
number of PLTs and allocate an array of Plt_Entry objects in the
Obj_Entry.  Instead of storing the Obj_Entry pointer in GOTs, store
the Plt_Entry pointer.

This currently lacks powerpc ELFv1 support.