sys/tools/syscalls: Support a separate abi_uintptr_t
Currently the code assumes that the substitution for uintptr_t is the
same as for intptr_t prefixed with a "u".  However, we'd like to support
per-ABI type name prefixes.

Introduce a separate abi_uintptr_t config key in order to support this.
If not set, it will default to abi_intptr_t with a "u" prefix.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK