libkern/strlen.c: make it work on CHERI
On CHERI we can't count on any particularly alignment of the front or
end of the capability and thus need to avoid reads which might land
outside the bounds in whole or in part until we make a final read
entierly outside the bounds in the case that the string is unterminated.

This implementation could be better in that it checks how much is left
of the capability on each loop rather than only doing so if the end is
under-aligned.

The non-CHERI case is unchanged.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK, DARPA, AFRL